feat: check if filename is str and file exists

+++ python-serialization/task_00_basic_serialization.py
@@ -4,19 +4,22 @@ Module that serializes dictionary into JSON file and
 deserialize JSON file to recreate a dictionary.
 """
 import pickle
+from os import path
 
 
 def serialize_and_save_to_file(data, filename):
-    try:
-        with open(filename, mode="wb") as a_file:
-            pickle.dump(data, a_file)
-    except Exception as error:
-        print(error)
+    if (isinstance(filename, str)) and path.exists(filename):
+        try:
+            with open(filename, mode="wb") as a_file:
+                pickle.dump(data, a_file)
+        except Exception as error:
+            print(error)
 
 
 def load_and_deserialize(filename):
-    try:
-        with open(filename, mode="rb") as a_file:
-            return pickle.load(a_file)
-    except Exception as error:
-        print(error)
+    if (isinstance(filename, str)) and path.exists(filename):
+        try:
+            with open(filename, mode="rb") as a_file:
+                return pickle.load(a_file)
+        except Exception as error:
+            print(error)
